Gemini is working, though responses py not
